Moved warnings into a separated component
